+++ examples/e2e/test/provider.spec.js
@@ -6,6 +6,7 @@ const { server, importData, animalRepository } = require("../provider.js")
 const path = require("path")
 
 server.listen(8081, () => {
+  importData()
   console.log("Animal Profile Service listening on http://localhost:8081")
 })