+++ examples/jest/__tests__/index.spec.js
@@ -1,35 +1,42 @@
 "use strict"
 
-const getMeDogs = require("../index").getMeDogs
+const { pactWith } = require("jest-pact")
+const { Matchers } = require("@pact-foundation/pact")
 
-describe("Dog's API", () => {
-  let url = "http://localhost"
+const { getMeDogs, getMeCats } = require("../index")
 
-  const EXPECTED_BODY = [
-    {
-      dog: 1,
-    },
-  ]
+pactWith({ consumer: "MyConsumer", provider: "MyProvider" }, provider => {
+  describe("Dogs API", () => {
+    const DOGS_DATA = [
+      {
+        dog: 1,
+      },
+    ]
 
-  describe("works", () => {
+    const dogsSuccessResponse = {
+      status: 200,
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: DOGS_DATA,
+    }
+
+    const dogsListRequest = {
+      uponReceiving: "a request for dogs",
+      withRequest: {
+        method: "GET",
+        path: "/dogs",
+        headers: {
+          Accept: "application/json",
+        },
+      },
+    }
+
     beforeEach(() => {
       const interaction = {
-        state: "i have a list of projects",
-        uponReceiving: "a request for projects",
-        withRequest: {
-          method: "GET",
-          path: "/dogs",
-          headers: {
-            Accept: "application/json",
-          },
-        },
-        willRespondWith: {
-          status: 200,
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: EXPECTED_BODY,
-        },
+        state: "i have a list of dogs",
+        ...dogsListRequest,
+        willRespondWith: dogsSuccessResponse,
       }
       return provider.addInteraction(interaction)
     })
@@ -37,15 +44,52 @@ describe("Dog's API", () => {
     // add expectations
     it("returns a sucessful body", () => {
       return getMeDogs({
-        url,
-        port,
+        url: provider.mockService.baseUrl,
+      }).then(dogs => {
+        expect(dogs).toEqual(DOGS_DATA)
       })
-        .then(response => {
-          expect(response.headers["content-type"]).toEqual("application/json")
-          expect(response.data).toEqual(EXPECTED_BODY)
-          expect(response.status).toEqual(200)
-        })
-        .then(() => provider.verify())
+    })
+  })
+
+  describe("Cats API", () => {
+    const CATS_DATA = [{ cat: 2 }, { cat: 3 }]
+
+    const catsSuccessResponse = {
+      status: 200,
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: CATS_DATA,
+    }
+
+    const catsListRequest = {
+      uponReceiving: "a request for cats with given catId",
+      withRequest: {
+        method: "GET",
+        path: "/cats",
+        query: {
+          "catId[]": Matchers.eachLike("1"),
+        },
+        headers: {
+          Accept: "application/json",
+        },
+      },
+    }
+
+    beforeEach(() => {
+      return provider.addInteraction({
+        state: "i have a list of cats",
+        ...catsListRequest,
+        willRespondWith: catsSuccessResponse,
+      })
+    })
+
+    it("returns a sucessful body", () => {
+      return getMeCats({
+        url: provider.mockService.baseUrl,
+      }).then(cats => {
+        expect(cats).toEqual(CATS_DATA)
+      })
     })
   })
 })