+++ examples/jest/index.js
@@ -4,24 +4,26 @@ const axios = require("axios")
 
 exports.getMeDogs = endpoint => {
   const url = endpoint.url
-  const port = endpoint.port
 
-  return axios.request({
-    method: "GET",
-    baseURL: `${url}:${port}`,
-    url: "/dogs",
-    headers: { Accept: "application/json" },
-  })
+  return axios
+    .request({
+      method: "GET",
+      baseURL: url,
+      url: "/dogs",
+      headers: { Accept: "application/json" },
+    })
+    .then(response => response.data)
 }
 
 exports.getMeCats = endpoint => {
   const url = endpoint.url
-  const port = endpoint.port
 
-  return axios.request({
-    method: "GET",
-    baseURL: `${url}:${port}`,
-    url: "/cats?catId[]=2&catId[]=3",
-    headers: { Accept: "application/json" },
-  })
+  return axios
+    .request({
+      method: "GET",
+      baseURL: url,
+      url: "/cats?catId[]=2&catId[]=3",
+      headers: { Accept: "application/json" },
+    })
+    .then(response => response.data)
 }