+++ examples/messages/provider/message-provider.spec.ts
@@ -15,12 +15,10 @@ describe("Message provider tests", () => {
         return Promise.resolve(`state set to create a dog`)
       },
     },
-    consumer: "MyJSMessageConsumer",
-    log: path.resolve(process.cwd(), "logs"),
     logLevel: "info",
     provider: "MyJSMessageProvider",
     providerVersion: "1.0.0",
-
+    consumer: "MyJSMessageConsumer",
     // For local validation
     // pactUrls: [path.resolve(process.cwd(), "pacts", "myjsmessageconsumer-myjsmessageprovider.json")],
     // Broker validation